rename codes for consistency
