+++ lst_scripts/hiperta_r0_to_dl1lstchain.py
@@ -7,7 +7,7 @@ import os
 import argparse
 from distutils.util import strtobool
 #from hipecta.programs.reorganize_dl1_files import reorganize_dl1
-from reorganize_dl1_output_files_hiperta_hipecta import reorganize_dl1
+from reorganize_dl1hiperta_to_dl1lstchain import reorganize_dl1
 
 
 parser = argparse.ArgumentParser(description="Run hiperta_r0_dl1 and reorganize_dl1_files")
